Experiencia de Aprendizaje 1 - Fundamentos de AI Generativa y Prompt Engineering/Clase 1.4/src/utils/mongodb.py
import os
from pymongo import MongoClient
from pymongo.collection import Collection
from dotenv import load_dotenv

load_dotenv()


class MongoDBClient:
    def __init__(self, db_name: str):
        self.client = MongoClient(os.getenv("MONGODB_CONNECTION_STRING"))
        self.db = self.client[db_name]

    def get_collection(self, collection_name: str) -> Collection:
        return self.db[collection_name]