Experiencia de Aprendizaje 1 - Fundamentos de AI Generativa y Prompt Engineering/Clase 1.4/requirements.txt
openai
pymongo
streamlit
markitdown[pdf]
gitpython
langchain
langchain-openai
langchain-mongodb
python-dotenv
langsmith
ragas
datasets